+++ src/Symfony/Component/ExpressionLanguage/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add support for null-coalescing unknown variables
  * Add support for comments using `/*` & `*/`
+ * Allow passing any iterable as `$providers` list to `ExpressionLanguage` constructor
 
 7.1
 ---